Add deque_at_wrapping()

+++ src/deque/ft_deque_at.c
@@ -21,11 +21,11 @@
 
 ////////////////////////////////////////////////////////////////////////////////
 
-void	*ft_deque_at(t_deque *deque, intptr_t index)
+void	*ft_deque_at(t_deque *deque, size_t index)
 {
 	size_t	wrapped_index;
 
-	wrapped_index = deque_get_wrapped_index(deque, index);
+	wrapped_index = deque_get_wrapped_index(deque, (intptr_t)index);
 	return (((t_u8 *)deque->data) + wrapped_index * deque->element_size);
 }
